shell: fix native data-layer build (async-trait/tokio deps, gc usize→u64)
cargo check on the Mac (the only host with the Apple/webkit toolchain) surfaced
what hawk can't compile: the HttpFetcher impl needs the async-trait macro in
scope, the manifest RwLock needs the tokio crate (Tauri doesn't re-export
tokio::sync), and Engine::gc returns usize while lv_gc returns u64. Add the two
deps + cast. cargo check now passes on macOS.

+++ src-tauri/src/sync.rs
@@ -211,7 +211,7 @@ pub(crate) async fn lv_sync_book(state: State<'_, LvState>, slug: String) -> Res
 #[tauri::command]
 pub(crate) async fn lv_gc(state: State<'_, LvState>) -> Result<u64, String> {
     let m = state.manifest.read().await;
-    Ok(state.engine.gc(&m).await)
+    Ok(state.engine.gc(&m).await as u64)
 }
 
 /// Initialise the data layer on the given app: build the state from the app data
